requirements.txt
# Core
ipython==7.13.0
Django==3.0.5
django-cors-headers==3.2.1
dj-database-url==0.5.0
django-debug-toolbar==2.2
mysqlclient==1.4.6
psycopg2==2.8.4
cx_Oracle==7.3.0
sql_server.pyodbc==1.0
python-dateutil==2.8.1
stringcase==1.2.0
unicodecsv==0.14.1
urllib3==1.25.8
numpy==1.18.2
pandas==1.0.3
Pillow==7.0.0
phonenumbers==8.12.1
python-magic==0.4.18
djangorestframework==3.11.0
django-oauth-toolkit==1.3.2
-e git://github.com/aschn/drf-tracking.git#egg=drf-tracking
django-import-export==2.0.2
tablib==1.1.0
odfpy==1.4.1
openpyxl==3.0.3
xlrd==1.2.0
xlwt==1.3.0
XlsxWriter==1.3.3
PyYAML==5.3.1
MarkupPy==1.14

# WebSockets
channels==3.0.3
channels-redis==3.2.0


# Server
gunicorn==20.0.4
gevent==1.4.0
greenlet==0.4.15

# Async queue
SQLAlchemy==1.3.15
celery==4.4.2
kombu==4.6.8
django-celery-beat==2.0.0
django-celery-results==1.2.1
redis==3.4.1

# Transport protocol
# requests==2.23.0
zeep==3.4.0

# Geolocation
geoip2==4.2.0
geopy==2.1.0
django-ipware==3.0.2

# scikit
scikit-learn==0.24.2

# nltk
nltk==3.6.2
